requirements.txt
numpy
wandb
tqdm
accelerate==1.4.0
codetiming==1.4.0
datasets==3.3.2
filelock==3.18.0
flash_attn==2.7.4.post1
liger_kernel==0.5.6
mathruler==0.1.0
mlflow==2.22.0
omegaconf==2.3.0
Pillow==11.2.1
psutil==6.1.0
PyYAML==6.0.2
qwen_vl_utils==0.0.11
swanlab==0.5.8
tensordict==0.5.0
torch==2.5.1
torchdata==0.11.0
vllm==0.7.3
